requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
python-telegram-bot==20.7
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
Pillow==10.1.0
yt-dlp==2023.12.30
flask==3.0.0
requests==2.31.0
psutil==5.9.6
humanize==4.9.0
asyncio==3.4.3
dnspython==2.4.2
